requirements.txt
google-auth-oauthlib
google-auth
requests
ultralytics
opencv-python